feature
Can scan for obsolete or expired download urls in settings.
Drivers not found have macro identified using a red "url" tag.
Macro names cannot have a comma nor a semicolon.
Phrases or words can be renamed when the macro is displayed.  This affected HP Smart which is now HP App